Fix Tupfiles with MALLOC symbols.
I had previously prefixed these with __UCLIBC__ since that's how I was parsing for macros, but then realized I would have to change a bunch of symbols like that, so I undid the patches and instead fixed it in libcpp. However, when I rebased the MALLOC symbol patch out of the way, I forgot to update these Tupfiles, and subsequently got confused.