Fix __h_errno_location?
I'm confused by this. I added a printf, and __h_errno_location is NULL. But the if(__h_errno_location != NULL) block still triggers. Inside that if I can print it again, and it's still NULL. But then when *(__h_errno_location()) triggers, it dies because it's NULL. Go figure.